refactor: navigation mobile 삭제, naviagtion 구조 수정
