+++ src/components/common/MainLogo/styled.ts
@@ -10,16 +10,10 @@ export const DaejinUnivLogoWrapper = styled.div`
   padding-bottom: 5px;
   display: flex;
   margin-left: 5px;
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
-    display: none;
-  }
 `;
 export const GdscBlogLogoWrapper = styled.div`
   display: flex;
   padding-bottom: 20px;
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
-    display: none;
-  }
 `;
 export const SubLogoWrapper = styled.div`
   align-items: center;
