--- src/components/common/Navigation/MobileNavigation/index.tsx
@@ -1,32 +1,0 @@
-import React from 'react';
-import { LogoWrapper } from '../DeskNavigation/styled';
-import {
-  MobileNavigationWrapper,
-  GdscLogoWrapper,
-  MobileVectorWrapper,
-  MobileNavigationDesign,
-} from './styled';
-import GdscLogo from '../../../../assets/GdscLogo';
-import Vector from '../../../../assets/Vector';
-import { useNavigate } from 'react-router';
-
-const MobileNavigation = () => {
-  const navigate = useNavigate();
-
-  return (
-    <MobileNavigationDesign>
-      <MobileNavigationWrapper>
-        <LogoWrapper onClick={() => navigate('/')}>
-          <GdscLogoWrapper>
-            <GdscLogo />
-          </GdscLogoWrapper>
-        </LogoWrapper>
-        <MobileVectorWrapper>
-          <Vector />
-        </MobileVectorWrapper>
-      </MobileNavigationWrapper>
-    </MobileNavigationDesign>
-  );
-};
-
-export default MobileNavigation;
