+++ src/components/common/Navigation/styled.ts
@@ -10,15 +10,12 @@ export const NavigationDesign = styled.nav`
   @media screen and (max-width: ${(props) => props.theme.windowSize.mobile}px) {
     height: 60px;
   }
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
-    display: none;
-  }
 `;
 export const NavigationWrapper = styled.div`
   display: flex;
   flex-direction: row;
-  width: 100%;
-  margin: auto;
+  width: 92%;
+  margin: 0 auto;
   height: 100%;
 `;
 export const NavigationInner = styled.div`
@@ -37,40 +34,33 @@ export const LogoWrapper = styled.div`
   display: flex;
   align-items: center;
   cursor: pointer;
+  margin-left: 50px;
+  @media (max-width: ${({ theme }) => theme.windowSize.tablet}px) {
+    display: none;
+  }
 `;
 export const DeskNavigationWrapper = styled.div`
   width: 100%;
   height: 100%;
   align-items: center;
+  justify-content: space-between;
   display: flex;
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
+  position: relative;
+`;
+export const VectorWrapper = styled.div`
+  @media (min-width: ${({ theme }) => theme.windowSize.tablet}px) {
     display: none;
   }
 `;
-export const BeforeMargin = styled.div`
+export const SearchInputWrapper = styled.div`
   display: flex;
-  width: 110px;
-  height: 100%;
-`;
-export const NavigationMargin = styled.div`
-  flex-grow: 1;
-  min-width: 20px;
-`;
-export const VectorWrapper = styled.div`
-  display: flex;
-  position: absolute;
-  right: 20px;
-  top: 9px;
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
-    & svg {
-      position: absolute;
-      top: -20px;
-      right: 0px;
-    }
-  }
+  position: relative;
   & svg {
     cursor: pointer;
-    & hover svg path {
+    position: absolute;
+    right: 20px;
+    top: 8px;
+    &:hover {
       fill: ${(props) => props.theme.color.grey300};
     }
   }
@@ -78,23 +68,37 @@ export const VectorWrapper = styled.div`
 export const SearchWrapper = styled.label`
   display: flex;
   margin-right: 40px;
-  position: relative;
+  align-items: center;
+  flex: 1;
+  justify-content: flex-end;
+  @media (max-width: ${({ theme }) => theme.windowSize.tablet}px) {
+    display: none;
+  }
 `;
 export const Search = styled.input`
   min-width: 400px;
   height: 30px;
-  font-size: ${(props) => props.theme.fontSize.body2};
+  font-size: ${({ theme }) => theme.fontSize.body2};
   border-radius: 50px;
-  border: 1px solid ${(props) => props.theme.color.grey300};
+  border: 1px solid ${({ theme }) => theme.color.grey300};
   padding-left: 20px;
   color: ${(props) => props.theme.color.grey900};
-  @media screen and (max-width: ${(props) => props.theme.windowSize.desk}px) {
+  @media (max-width: ${({ theme }) => theme.windowSize.desk}px) {
     min-width: 280px;
   }
-  @media screen and (max-width: ${(props) => props.theme.windowSize.tablet}px) {
+  @media (max-width: ${({ theme }) => theme.windowSize.tablet}px) {
     display: none;
   }
   &::placeholder {
-    color: ${(props) => props.theme.color.grey300};
+    color: ${({ theme }) => theme.color.grey300};
   }
+`;
+
+export const GdscLogoWrapper = styled.div`
+  & svg {
+    width: 40px;
+    height: 20px;
+    @media(min-width:  ${({ theme }) => theme.windowSize.tablet}px) {
+      display: none;
+    }
 `;
